EarlyAlertsAPI/pyproject.toml
[project]
name = "earlyalertsapi"
version = "0.1.0"
description = "Early Alerts API – Module 2c operational alert backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "sqlalchemy>=2.0",
    "httpx>=0.28",
    "typer>=0.15",
    "dlt[duckdb]>=1.6",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "ruff>=0.9",
    "radon>=6.0",
]

[project.scripts]
early-alerts = "app.backend.cli:app"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
]
